pyproject.toml
[tool.ruff]
line-length = 120
target-version = "py310"

[project]
name = "chachi_bt_finance"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "openai==1.60.2",
    "langchain==0.3.17",
    "streamlit==1.41.1",
    "ruff==0.9.3",
    "yfinance==0.2.52",
    "pandas==2.2.3",
    "numpy==1.26.4",
    "matplotlib==3.10.0",
    "seaborn==0.13.2",
    "python-dotenv==1.0.1",
    "pydantic==2.10.6",
    "langchain-community==0.3.16"
]
